Add files via upload
Programming Exercise 1 and 2 involves creating Recursion code to draw a specific shape.

This file does not have that Recursion code for Programming Exercise 1. However, it contains a For Loop you can use to get started, as it does draw the example shape seen in PE 1. It also asks you how large you want the "table" to be(or the shape, the file is modified from Bryson's multiplication table).